fix: production Docker hardening — restart policy, BigInteger, non-root (#342)
- Add restart: unless-stopped to both services in docker-compose.prod.yaml
- Change volume column from Integer to BigInteger (model + migration 0008)
- Add non-root appuser to production Dockerfile
- Add --no-cache-dir to pip install in production Dockerfile
- Add backend healthcheck to docker-compose.yaml and docker-compose.prod.yaml
- Bind PostgreSQL to 127.0.0.1 in dev docker-compose.yaml
- Replace hardcoded DB credentials with placeholder in alembic.ini
- Remove dead nginx prod stage from frontend/Dockerfile

+++ backend/app/models/price.py
@@ -1,6 +1,6 @@
 from datetime import date
 
-from sqlalchemy import Date, ForeignKey, Integer, Numeric, UniqueConstraint
+from sqlalchemy import BigInteger, Date, ForeignKey, Numeric, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.database import Base
@@ -17,7 +17,7 @@ class PriceHistory(Base):
     high: Mapped[float] = mapped_column(Numeric(12, 4))
     low: Mapped[float] = mapped_column(Numeric(12, 4))
     close: Mapped[float] = mapped_column(Numeric(12, 4))
-    volume: Mapped[int] = mapped_column(Integer)
+    volume: Mapped[int] = mapped_column(BigInteger)
 
     asset: Mapped["Asset"] = relationship(back_populates="prices")
